[fix]: tabs for news alert and papers
