+++ components/NewsPage.tsx
@@ -1,7 +1,7 @@
 /** @format */
 "use client";
 
-import React, { useEffect, useState, useRef } from "react";
+import React, { useEffect, useState, useRef, Suspense } from "react";
 import {
     Box,
     Container,
@@ -16,10 +16,13 @@ import {
     LinkOverlay,
     Flex,
     useColorModeValue,
+    Tabs,
+    TabList,
+    Tab,
+    TabPanels,
+    TabPanel,
 } from "@chakra-ui/react";
-import { ExternalLinkIcon } from "@chakra-ui/icons";
-import { useRouter, useSearchParams } from "next/navigation";
-import newsData from "@/data/news.json";
+import NewsContainer from "./NewsContainer";
 
 interface NewsItem {
     date: string;
@@ -39,84 +42,27 @@ const News: React.FC = () => {
     const cardHeading = useColorModeValue("yellow.600", "whiteAlpha.800")
     const textCol = useColorModeValue("gray.500", "gray.400")
 
-    const router = useRouter();
-    const searchParams = useSearchParams();
-    const [searchTerm, setSearchTerm] = useState<string>(searchParams.get("q") || "");
-    const [filteredNews, setFilteredNews] = useState<NewsItem[]>([]);
-    const [debouncing, setDebouncing] = useState<boolean>(false);
-    const [currentPage, setCurrentPage] = useState<number>(
-        parseInt(searchParams.get("page") || "1", 10)
-    );
-    const newsPerPage = 10;
-    const debounceTimer = useRef<NodeJS.Timeout | null>(null);
+    const [activeTab, setActiveTab] = useState(0);
 
-    // Sort news by most recent date
-    const sortedNews: NewsItem[] = newsData.sort(
-        (a, b) => new Date(b.date).getTime() - new Date(a.date).getTime()
-    );
+    // Function to handle tab change and update the active tab
+    const handleTabChange = (index: number) => {
+        setActiveTab(index);
+    };
 
-    // Filter news based on search term with debouncing
+    // Check URL hash and query parameters to set the active tab
     useEffect(() => {
-        if (debounceTimer.current) {
-            clearTimeout(debounceTimer.current);
+        document.title = "News - Artificial Intelligence in Transportation Lab (AIT Lab)";
+        const hash = window.location.hash;
+
+        // Handle tab selection based on URL hash
+        if (hash === "#alert") {
+            setActiveTab(0);
+        } else if (hash === "#papers") {
+            setActiveTab(1);
+        } else {
+            setActiveTab(0);
         }
-
-        setDebouncing(true);
-
-        debounceTimer.current = setTimeout(() => {
-            const filtered = sortedNews.filter((news) =>
-                news.title.toLowerCase().includes(searchTerm.toLowerCase())
-            );
-
-            setFilteredNews(filtered);
-            setCurrentPage(1); // Reset to the first page
-            setDebouncing(false);
-
-            // Update the URL with the search query and reset page to 1
-            const params = new URLSearchParams();
-            if (searchTerm) params.set("q", searchTerm);
-            params.set("page", "1");
-            router.replace(`?${params.toString()}`);
-        }, 500);
-
-        // Clean up debounce timer on unmount
-        return () => {
-            if (debounceTimer.current) {
-                clearTimeout(debounceTimer.current);
-            }
-        };
-    }, [searchTerm, sortedNews, router]);
-
-    // Load search term and page from URL on initial render
-    useEffect(() => {
-        const query = searchParams.get("q");
-        const page = parseInt(searchParams.get("page") || "1", 10);
-
-        if (query) setSearchTerm(query);
-        if (page) setCurrentPage(page);
-    }, [searchParams]);
-
-    // Pagination Logic
-    const indexOfLastNews = currentPage * newsPerPage;
-    const indexOfFirstNews = indexOfLastNews - newsPerPage;
-    const currentNews = filteredNews.slice(indexOfFirstNews, indexOfLastNews);
-    const totalPages = Math.ceil(filteredNews.length / newsPerPage);
-
-    const handlePageChange = (page: number) => {
-        setCurrentPage(page);
-
-        window.scrollTo({ top: 0, behavior: "smooth" });
-
-        // Update the URL with the new page number
-        const params = new URLSearchParams();
-        if (searchTerm) params.set("q", searchTerm);
-        params.set("page", page.toString());
-        router.replace(`?${params.toString()}`);
-    };
-
-    const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setSearchTerm(e.target.value);
-    };
+    }, []);
 
     return (
         <Box py={20} minH={"100%"} bgColor={bgColor}>
@@ -125,76 +71,33 @@ const News: React.FC = () => {
                     News & Updates
                 </Heading>
 
-                {/* Search input */}
-                <Stack mb={6} spacing={4}>
-                    <Input
-                        placeholder="Search news by title"
-                        value={searchTerm}
-                        onChange={handleSearch}
-                        bg={inputBg}
-                        borderColor={inputBorder}
-                        _placeholder={{ color: placeHolderColor }}
-                    />
-                </Stack>
 
-                {debouncing ? (
-                    <Center py={10}>
-                        <Spinner size="xl" color="yellow.500" />
-                    </Center>
-                ) : currentNews.length > 0 ? (
-                    <Stack spacing={3}>
-                        {currentNews.map((news, index) => (
-                            <LinkBox
-                                key={index}
-                                as="article"
-                                p={5}
-                                shadow="md"
-                                borderWidth="1px"
-                                borderRadius="md"
-                                bg={cardBgColor}
-                                _hover={{ shadow: "lg", transform: "translateY(-5px)" }}
-                                transition="all 0.3s ease"
-                                cursor={news.link ? "pointer" : "default"}
-                            >
-                                <Flex justify="space-between" align="center" mb={2}>
-                                    <Text fontWeight="bold" color={textCol} fontSize="md">
-                                        {new Date(news.date).toLocaleDateString("en-US", {
-                                            year: "numeric",
-                                            month: "long",
-                                            day: "numeric",
-                                        })} :: {news.title}
-                                    </Text>
-                                </Flex>
+                <Tabs index={activeTab} onChange={handleTabChange} colorScheme="yellow">
+                    <TabList mb={4}>
+                        <Tab>Alerts</Tab>
+                        <Tab>Papers</Tab>
+                    </TabList>
 
-                                <Text color={cardHeading} fontSize="lg" fontWeight="bold">
-                                    {news.link ? (
-                                        <LinkOverlay href={news.link} isExternal>
-                                            {news.description} <ExternalLinkIcon mx="2px" />
-                                        </LinkOverlay>
-                                    ) : (
-                                        news.description
-                                    )}
-                                </Text>
-                                {/* <Text color="gray.600">{news.description}</Text> */}
-                            </LinkBox>
-                        ))}
-                    </Stack>
-                ) : (
-                    <Text>No news found</Text>
-                )}
-
-                {/* Pagination Controls */}
-                <Stack direction="row" justify="center" mt={8}>
-                    {currentPage > 1 && (
-                        <Button onClick={() => handlePageChange(currentPage - 1)} _hover={{ color: "white", backgroundColor: "primary" }}>Previous</Button>
-                    )}
-                    <Center>
-                        Page {currentPage} of {totalPages}
-                    </Center>
-                    {currentPage < totalPages && (
-                        <Button onClick={() => handlePageChange(currentPage + 1)} _hover={{ color: "white", backgroundColor: "primary" }}>Next</Button>
-                    )}
-                </Stack>
+                    <TabPanels>
+                        <TabPanel>
+                            {activeTab === 0 && (
+                                <Suspense fallback={<Spinner size="xl" color="yellow.500" />}>
+                                    {/* Assign a unique key based on the active tab */}
+                                    <NewsContainer key={`tab-${activeTab}`} type="alert"
+                                    />
+                                </Suspense>
+                            )}
+                        </TabPanel>
+                        <TabPanel>
+                            {activeTab === 1 && (
+                                <Suspense fallback={<Spinner size="xl" color="yellow.500" />}>
+                                    {/* Assign a unique key based on the active tab */}
+                                    <NewsContainer key={`tab-${activeTab}`} type="all" />
+                                </Suspense>
+                            )}
+                        </TabPanel>
+                    </TabPanels>
+                </Tabs>
             </Container>
         </Box>
     );
